Add a basic project skeleton
This sets up the basic skeleton of the project to build documentation, run
tests, lint the codebase, and construct a code coverage report. We'll need to
extend this to handle publishing the package and the documentation, but this
should be enough of a framework to develop the actual library.

Merges #1