Add a scope module with a Scope class and related utilities
This adds a `dysco.scope` module with a `Scope` class and a handful of
utilities for finding scopes. I also ditched Python 3.6 for the time being
because it doesn't support `annotations` from `__future__`, and it would add a
lot of boilerplate to handle the types without that. I might add it in later,
but I want to get a reasonable first version done before that.

Merges #2