Add PyPy 3.6 as a CI target
This adds PyPy 3.6 as a testing target on CI and updates the trove classifiers.
The removal of mypy is a bit hacky, but it's required because `typed_ast` isn't
compatible. It was also an issue that `poetry install` ran very slowly even
when no dependencies actually need to be updated. To work around this, I stored
an md5sum in `poetry.lock.md5`, included it in the cache, and skipped running
`poetry install` when the md5sum hasn't changed.

Merges #24