Use invalid variable name strings instead of typles as f_locals keys
We were previously using a tuple containing the scope name as the key in
`frame.f_locals`, but it's simpler and more direct to just use a string with an
invalid variable name. This PR switches the code to use `<` and `>` in the
variable name to prevent potential collisions.

Merges #25